requirements.txt
streamlit
speechrecognition
pyttsx3
pyaudio
requests
python-dotenv